+++ plugins/joy_to_twist/JoyToTwist.hh
@@ -80,7 +80,7 @@ namespace ignition
       public: virtual ~JoyToTwist();
 
       // Documentation inherited
-      public: virtual void Load(
+      public: virtual bool Load(
                   const tinyxml2::XMLElement *_elem) override final;
 
       /// \brief Run the plugin