+++ src/Manager.cc
@@ -14,7 +14,11 @@
  * limitations under the License.
  *
 */
+
+#include <fcntl.h>
+#include <semaphore.h>
 #include <signal.h>
+#include <sys/stat.h>
 #include <sys/wait.h>
 #include <tinyxml2.h>
 #include <unistd.h>
@@ -23,6 +27,7 @@
 #include <list>
 #include <mutex>
 #include <numeric>
+#include <queue>
 #include <string>
 #include <thread>
 #include <unordered_set>
@@ -35,9 +40,14 @@
 
 #include "ignition/launch/config.hh"
 #include "ignition/launch/Plugin.hh"
+
 #include "Manager.hh"
+#include "vendor/backward.hpp"
 
 using namespace ignition::launch;
+using namespace std::chrono_literals;
+
+static constexpr const char* kSemaphoreName = "/child_semaphore";
 
 /// \brief A class to encapsulate an executable (program) to run.
 class Executable
@@ -86,6 +96,9 @@ class ignition::launch::ManagerPrivate
   /// \brief Constructor.
   public: ManagerPrivate();
 
+  /// \brief Destructor.
+  public: ~ManagerPrivate();
+
   /// \brief Parse a configuration string.
   /// \param[in] _string XML configuration string.
   /// \return True on success.
@@ -133,6 +146,12 @@ class ignition::launch::ManagerPrivate
   /// \param[in] _sig The signal
   private: static void OnSigChild(int _sig);
 
+  /// \brief Start the worker thread to service stopped children.
+  public: void StartWorkerThread();
+
+  /// \brief Thread to handle restarting stopped children.
+  private: void RestartLoop();
+
   /// \brief Parse executable configurations.
   /// \param[in] _elem XML element that contains an <executable>
   private: void ParseExecutables(const tinyxml2::XMLElement *_elem);
@@ -153,6 +172,15 @@ class ignition::launch::ManagerPrivate
   /// \brief A list of executables that are running, or have been run.
   public: std::list<Executable> executables;
 
+  /// \brief A list of children that were stopped to attempt restarts
+  public: std::queue<pid_t> stoppedChildren;
+
+  /// \brief Semaphore to prevent restartThread from being a spinlock
+  private: sem_t *stoppedChildSem;
+
+  /// \brief Thread containing the restart loop
+  private: std::thread restartThread;
+
   /// \brief All the plugins
   public: std::unordered_set<launch::PluginPtr> plugins;
 
@@ -169,13 +197,16 @@ class ignition::launch::ManagerPrivate
   public: std::mutex runMutex;
 
   /// \brief True if running.
-  public: bool running = false;
+  public: std::atomic<bool> running = false;
 
   /// \brief True indicates that this process is the master (main) process.
   public: bool master = false;
 
   /// \brief Our signal handler.
   public: std::unique_ptr<common::SignalHandler> sigHandler = nullptr;
+
+  /// \brief Backward signal handler
+  public: std::unique_ptr<backward::SignalHandling> backward = nullptr;
 
   /// \brief Top level environment variables.
   public: std::list<std::string> envs;
@@ -229,6 +260,9 @@ bool Manager::RunConfig(const std::string &_config)
   this->dataPtr->running = !this->dataPtr->executables.empty() ||
                            !this->dataPtr->plugins.empty();
 
+  // Start thread to service child signals.
+  this->dataPtr->StartWorkerThread();
+
   // Wait for a shutdown event, or for all the executables to quit.
   while (this->dataPtr->running && (!this->dataPtr->executables.empty() ||
                                     !this->dataPtr->plugins.empty()))
@@ -260,74 +294,182 @@ ManagerPrivate::ManagerPrivate()
   this->sigHandler->AddCallback(
       std::bind(&ManagerPrivate::OnSigIntTerm, this, std::placeholders::_1));
 
+  // Initialize semaphore
+  this->stoppedChildSem = sem_open(kSemaphoreName, O_CREAT, 0644, 1);
+  if (this->stoppedChildSem == SEM_FAILED) {
+    ignerr << "Error initializing semaphore: " << strerror(errno) << std::endl;
+  }
+
   // Register a signal handler to capture child process death events.
   if (signal(SIGCHLD, ManagerPrivate::OnSigChild) == SIG_ERR)
     ignerr << "signal(2) failed while setting up for SIGCHLD" << std::endl;
+
+  // Register backward signal handler for other signals
+  std::vector<int> signals = {
+    SIGABRT,    // Abort signal from abort(3)
+    SIGBUS,     // Bus error (bad memory access)
+    SIGFPE,     // Floating point exception
+    SIGILL,     // Illegal Instruction
+    SIGIOT,     // IOT trap. A synonym for SIGABRT
+    // SIGQUIT, // Quit from keyboard
+    SIGSEGV,    // Invalid memory reference
+    SIGSYS,     // Bad argument to routine (SVr4)
+    SIGTRAP,    // Trace/breakpoint trap
+    SIGXCPU,    // CPU time limit exceeded (4.2BSD)
+    SIGXFSZ,    // File size limit exceeded (4.2BSD)
+  };
+
+  this->backward = std::make_unique<backward::SignalHandling>(signals);
+}
+
+/////////////////////////////////////////////////
+ManagerPrivate::~ManagerPrivate()
+{
+  if (this->master)
+  {
+    if (sem_close(this->stoppedChildSem) == -1)
+    {
+      ignerr << "Failed to close semaphore: " << strerror(errno) << std::endl;
+    }
+
+    if (sem_unlink(kSemaphoreName) == -1)
+    {
+      ignerr << "Failed to unlink semaphore: " << strerror(errno) << std::endl;
+    }
+  }
 }
 
 /////////////////////////////////////////////////
 bool ManagerPrivate::Stop()
 {
-  std::lock_guard<std::mutex> lock(this->runMutex);
-
-  if (this->running)
+  if (this->runMutex.try_lock())
   {
-    this->running = false;
+    if (this->running)
+    {
+      this->running = false;
+    }
+    this->runMutex.unlock();
     this->runCondition.notify_all();
-  }
 
+    // Signal the restart thread to stop
+    sem_post(this->stoppedChildSem);
+    if (this->restartThread.joinable())
+      this->restartThread.join();
+  }
   return this->running;
 }
 
 /////////////////////////////////////////////////
 void ManagerPrivate::OnSigIntTerm(int _sig)
 {
-  igndbg << "Received signal[" << _sig  << "]\n";
+  igndbg << "OnSigIntTerm Received signal[" << _sig  << "]\n";
   this->Stop();
 }
 
 /////////////////////////////////////////////////
 void ManagerPrivate::OnSigChild(int _sig)
 {
-  igndbg << "Received signal[" << _sig  << "]\n";
+  // This is a signal handler, so be careful not to do any operations
+  // that you are not allowed to do.
+  // Ref: http://man7.org/linux/man-pages/man7/signal-safety.7.html
+  (void) _sig;  // Commenting _sig above confuses codecheck
   pid_t p;
   int status;
 
-  // Executable to restart
-  Executable restartExec;
-
+  // Retreive the stopped child's PID, append, and signal the consumer.
   if ((p = waitpid(-1, &status, WNOHANG)) != -1)
   {
-    std::lock_guard<std::mutex> mutex(myself->executablesMutex);
+    myself->stoppedChildren.push(p);
+    sem_post(myself->stoppedChildSem);
+  }
+}
 
-    // Find the executable
-    for (std::list<Executable>::iterator iter = myself->executables.begin();
-         iter != myself->executables.end(); ++iter)
+/////////////////////////////////////////////////
+void ManagerPrivate::StartWorkerThread()
+{
+  // Spawn thread to monitor restarting children.
+  this->restartThread = std::thread(&ManagerPrivate::RestartLoop, this);
+}
+
+/////////////////////////////////////////////////
+void ManagerPrivate::RestartLoop()
+{
+  sigset_t chldmask;
+
+  if ((sigemptyset(&chldmask) == -1) || (sigaddset(&chldmask, SIGCHLD) == -1))
+  {
+    ignerr << "Failed to initialize signal mask: "
+      << strerror(errno) << std::endl;
+    return;
+  }
+
+  while (this->running)
+  {
+    // Wait for the signal handler to signal that a child has exited.
+    int s = sem_wait(this->stoppedChildSem);
+    if (s == -1)
     {
-      if (iter->pid == p)
+      if (errno != EINTR)
+        ignwarn << "sem_wait error: "
+          << strerror(errno) << std::endl;
+      continue;
+    }
+
+    // Block SIGCHLD while consuming queue.
+    if (sigprocmask(SIG_BLOCK, &chldmask, NULL) == -1)
+    {
+      ignerr << "Failed to setup block for SIGCHLD: "
+        << strerror(errno) << std::endl;
+      continue;
+    }
+
+    // Consume stopped children from the queue.
+    while (!this->stoppedChildren.empty())
+    {
+      std::lock_guard<std::mutex> mutex(this->executablesMutex);
+
+      // Executable to restart
+      Executable restartExec;
+
+      pid_t p = this->stoppedChildren.front();
+      this->stoppedChildren.pop();
+
+      // Find the executable
+      for (std::list<Executable>::iterator iter = this->executables.begin();
+           iter != this->executables.end(); ++iter)
       {
-        igndbg << "Death of process[" << p << "] with name["
-               << iter->name << "].\n";
+        if (iter->pid == p)
+        {
+          igndbg << "Death of process[" << p << "] with name["
+                 << iter->name << "].\n";
 
+          // Restart if autoRestart is enabled
+          if (iter->autoRestart)
+            restartExec = *iter;
 
-        // Restart if autoRestart is enabled
-        if (iter->autoRestart)
-          restartExec = *iter;
-
-        myself->executables.erase(iter);
-        break;
+          this->executables.erase(iter);
+          break;
+        }
       }
+
+      if (!restartExec.name.empty() && !restartExec.command.empty())
+      {
+        igndbg << "Restarting process with name[" << restartExec.name << "]\n";
+        this->RunExecutable(restartExec);
+      }
+
+      this->runCondition.notify_all();
+    }
+
+    // Unblock SIGCHLD
+    if (sigprocmask(SIG_UNBLOCK, &chldmask, NULL) == -1)
+    {
+      ignerr << "Failed to unblock SIGCHLD: " << strerror(errno) << std::endl;
+      continue;
     }
   }
-
-  if (!restartExec.name.empty() && !restartExec.command.empty())
-  {
-    igndbg << "Restarting process with name[" << restartExec.name << "]\n";
-    myself->RunExecutable(restartExec);
-  }
-
-  myself->runCondition.notify_all();
 }
+
 
 /////////////////////////////////////////////////
 bool ManagerPrivate::ParseConfig(const std::string &_config)
@@ -563,6 +705,13 @@ std::list<std::string> ManagerPrivate::ParseEnvs(
     {
       std::string name = nameElem->GetText();
       std::string value = valueElem->GetText();
+
+      // Expand env var contents, such as $LD_LIBRARY_PATH
+      if (!value.empty() && value.at(0) == '$')
+      {
+        ignition::common::env(value.substr(1), value);
+      }
+
       result.push_back(name + "=" + value);
     }
 
@@ -596,6 +745,15 @@ void ManagerPrivate::LoadPlugin(const tinyxml2::XMLElement *_elem)
     return;
   }
 
+  if (name == "ignition::launch::GazeboServer")
+  {
+    setenv("RMT_PORT", "1500", 1);
+  }
+  else if (name == "ignition::launch::GazeboGui")
+  {
+    setenv("RMT_PORT", "1501", 1);
+  }
+
   ignition::common::SystemPaths systemPaths;
   systemPaths.SetPluginPathEnv("IGN_LAUNCH_PLUGIN_PATH");
   systemPaths.AddPluginPaths(IGNITION_LAUNCH_PLUGIN_INSTALL_PATH);
@@ -612,7 +770,12 @@ void ManagerPrivate::LoadPlugin(const tinyxml2::XMLElement *_elem)
   ignition::common::env(IGN_HOMEDIR, homePath);
   systemPaths.AddPluginPaths(homePath + "/.ignition/gazebo/plugins");
 
-  std::string pathToLib = systemPaths.FindSharedLibrary(file);
+  std::string pathToLib;
+  if (common::exists(file))
+    pathToLib = file;
+  else
+    pathToLib = systemPaths.FindSharedLibrary(file);
+
   if (pathToLib.empty())
   {
     ignerr << "Failed to find the path to library[" << file << "]. "
@@ -647,8 +810,8 @@ void ManagerPrivate::LoadPlugin(const tinyxml2::XMLElement *_elem)
     << "] File[" << file << "]" << std::endl;
 
   PluginPtr plugin = loader.Instantiate(name);
-  plugin->QueryInterface<Plugin>()->Load(_elem);
-  this->plugins.insert(plugin);
+  if (plugin->QueryInterface<Plugin>()->Load(_elem))
+    this->plugins.insert(plugin);
 }
 
 //////////////////////////////////////////////////
@@ -685,6 +848,7 @@ void ManagerPrivate::ParseExecutableWrappers(
         this->plugins.clear();
         this->wrappedPlugins.clear();
         this->sigHandler.reset();
+        this->backward.reset();
         this->executables.clear();
         this->LoadPlugin(pluginElem);
         return;