+++ Changelog.md
@@ -1,4 +1,71 @@
+## Ignition Launch 1.x
+
+### Ignition Launch 1.X.X
+
+### Ignition Launch 1.2.3 (2019-09-09)
+
+1. Add QML import path to IgnGazebo modules
+    * [Pull Request 42](https://bitbucket.org/ignitionrobotics/ign-launch/pull-requests/42)
+
+### Ignition Launch 1.2.2
+
+1. Add support for console logging.
+    * [Pull Request 40](https://bitbucket.org/ignitionrobotics/ign-launch/pull-requests/40)
+
+1. Support setting custom window icons.
+    * [Pull Request 39](https://bitbucket.org/ignitionrobotics/ign-launch/pull-requests/39)
+
+### Ignition Launch 1.2.1
+
+1. Eliminate potential deadlock from SIGCHLD signal handler
+    * [Pull Request 36](https://bitbucket.org/ignitionrobotics/ign-launch/pull-requests/36)
+
+### Ignition Launch 1.2.0
+
+1. Support for custom random seed in the GazeboServer plugin.
+    * [Pull Request 33](https://bitbucket.org/ignitionrobotics/ign-launch/pull-requests/33)
+
+1. Allow specifying a custom window title
+    * [Pull Request 32](https://bitbucket.org/ignitionrobotics/ign-launch/pull-requests/32)
+
+### Ignition Launch 1.1.0 (2019-06-11)
+
+1. Added command line parameters of the form <name>:=<value>. These parameters are passed to ERB.
+    * [Pull Request 27](https://bitbucket.org/ignitionrobotics/ign-launch/pull-requests/27)
+
+### Ignition Launch 1.0.1 (2019-05-22)
+
+1. Fix GazeboFactory set performer topic
+    * [Pull Request 26](https://bitbucket.org/ignitionrobotics/ign-launch/pull-requests/26)
+
+### Ignition Launch 1.0.0 (2019-05-21)
+
+1. Enable logging with `<record>` tag in ign launch file.
+    * [Pull Request 23](https://bitbucket.org/ignitionrobotics/ign-launch/pull-requests/23)
+
+1. Fix gazebo gui to use GuiRunner.
+    * [Pull Request 22](https://bitbucket.org/ignitionrobotics/ign-launch/pull-requests/22)
+
+1. Fix parsing positional argument to ign tool.
+    * [Pull Request 21](https://bitbucket.org/ignitionrobotics/ign-launch/pull-requests/21)
+
+1. Added world stats to Websocket server.
+    * [Pull Request 20](https://bitbucket.org/ignitionrobotics/ign-launch/pull-requests/20)
+
+1. Change project name to ignition-launch1.
+    * [Pull Request 19](https://bitbucket.org/ignitionrobotics/ign-launch/pull-requests/19)
+
+1. Depend on gazebo2, gui2, msgs4, transport7.
+    * [Pull Request 17](https://bitbucket.org/ignitionrobotics/ign-launch/pull-requests/17)
+
+1. Added Websocket server.
+    * [Pull Request 15](https://bitbucket.org/ignitionrobotics/ign-launch/pull-requests/15)
+    * [Pull Request 18](https://bitbucket.org/ignitionrobotics/ign-launch/pull-requests/18)
+
 ## Ignition Launch 0.x
 
-### Ignition Launch 0.1.0
+1. Install examples folder.
+    * [Pull Request 14](https://bitbucket.org/ignitionrobotics/ign-launch/pull-requests/14)
+
+### Ignition Launch 0.2.0